chimera_api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
supabase==2.10.0
google-generativeai==0.8.3
google-api-python-client==2.157.0
google-auth==2.37.0
pyjwt==2.10.1
cryptography==44.0.0
pytest==8.3.4
pytest-asyncio==0.25.2